+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - small alekseon logo
 - cache tags
-- getRecordCollection() method in Form class
+- getRecordCollection() and getRecordById() methods in Form class
 - filter,sort,select on form records collection by field identifier
 - getData from form record by field identifier